adding aclSources for relays initial UI preview, api support, database migration

+++ pages/api/relay/[id]/allowlistpubkeys.ts
@@ -96,7 +96,7 @@ export default async function handle(req: any, res: any) {
             return;
         }
         // delete all lists with this reason..
-        if (listId != "") {
+        if (listId != "" && listId != "all") {
             await prisma.listEntryPubkey.deleteMany({
                 where: {
                     AND: [{ AllowListId: allow_list.id }, { reason: { contains: listId }}],
